Moved streamDistribute out into its own file
