+++ hail/src/main/scala/is/hail/expr/ir/EmitStreamDistribute.scala
@@ -1,0 +1,279 @@
+package is.hail.expr.ir
+
+import is.hail.annotations.Region
+import is.hail.asm4s.{Code, Value, const}
+import is.hail.expr.ir.functions.MathFunctions
+import is.hail.expr.ir.orderings.StructOrdering
+import is.hail.expr.ir.streams.EmitStream
+import is.hail.io.OutputBuffer
+import is.hail.types.VirtualTypeWithReq
+import is.hail.types.physical.{PBooleanRequired, PCanonicalArray, PCanonicalInterval, PCanonicalStringRequired, PCanonicalStruct, PContainer, PInt32Required}
+import is.hail.types.physical.stypes.EmitType
+import is.hail.types.physical.stypes.concrete.{SIndexablePointer, SIndexablePointerCode, SJavaString, SStackStruct, SStackStructCode}
+import is.hail.types.physical.stypes.interfaces.{SBaseStruct, SIndexableCode, SIndexableValue, SStreamCode, SStringCode, primitive}
+import is.hail.types.physical.stypes.primitives.{SBooleanCode, SInt32, SInt32Code}
+import is.hail.types.virtual.{TArray, TStruct}
+import is.hail.utils._
+import is.hail.asm4s._
+
+object EmitStreamDistribute {
+
+  def emit(emitter: Emit[_], ir: StreamDistribute, cb: EmitCodeBuilder, region: Value[Region], env: EmitEnv, typeWithReqx: IR => VirtualTypeWithReq, container: Option[AggContainer], loopEnv: Option[Env[LoopRef]]): IEmitCode = {
+    val (child, pivots, path, spec) = (ir.child, ir.pivots, ir.path, ir.spec)
+    def emitI(ir: IR) = emitter.emitI(ir, cb, region, env, container, loopEnv)
+    def emitStream(ir: IR, cb: EmitCodeBuilder, outerRegion: Value[Region]): IEmitCode =
+      EmitStream.produce(emitter, ir, cb, outerRegion, env, container)
+    val mb = cb.emb
+
+    emitI(path).flatMap(cb) { case pathCode: SStringCode =>
+      val pathVal = pathCode.memoize(cb, "stream_dist_output_path")
+
+      emitI(pivots).flatMap(cb) { case pivotsCode: SIndexableCode =>
+        emitStream(child, cb, region).map(cb) { case childStream: SStreamCode =>
+          val requestedSplittersVal = pivotsCode.memoize(cb, "stream_dist_pivots")
+
+          val keyType = requestedSplittersVal.st.elementType.asInstanceOf[SBaseStruct]
+          val keyPType = typeWithReqx(pivots).canonicalPType.asInstanceOf[PContainer].elementType
+          val keyFieldNames = keyType.virtualType.fields.map(_.name)
+
+          def compare(cb: EmitCodeBuilder, lelt: EmitValue, relt: EmitValue): Code[Int] = {
+            val lhs = EmitCode.fromI(mb)(cb => lelt.toI(cb).map(cb)(_.asBaseStruct.subset(keyFieldNames: _*)))
+            val rhs = EmitCode.fromI(mb)(cb => relt.toI(cb).map(cb)(_.asBaseStruct.subset(keyFieldNames: _*)))
+            StructOrdering.make(lhs.st.asInstanceOf[SBaseStruct], rhs.st.asInstanceOf[SBaseStruct],
+              cb.emb.ecb, missingFieldsEqual = true)
+              .compare(cb, lhs, rhs, missingEqual = true)
+          }
+
+          def equal(cb: EmitCodeBuilder, lelt: EmitValue, relt: EmitValue): Code[Boolean] = compare(cb, lelt, relt) ceq 0
+
+          def lessThan(cb: EmitCodeBuilder, lelt: EmitValue, relt: EmitValue): Code[Boolean] = compare(cb, lelt, relt) < 0
+
+          val filledInTreeSize = Code.invokeScalaObject1[Int, Int](MathFunctions.getClass, "roundToNextPowerOf2", requestedSplittersVal.loadLength() + 1)
+          val treeHeight: Value[Int] = cb.newLocal[Int]("stream_dist_tree_height", Code.invokeScalaObject1[Int, Int](MathFunctions.getClass, "log2", filledInTreeSize))
+
+          val numberOfBuckets = cb.newLocal[Int]("stream_dist_number_of_buckets", const(1) << (treeHeight + 1))
+
+          val paddedSplittersSize = cb.newLocal[Int]("stream_dist_padded_splitter_size", numberOfBuckets / 2)
+
+          // Copy each unique splitter into array. If it is seen twice, set a boolean in a parallel array for that
+          // splitter, so we know what identity buckets to make later
+
+          val paddedSplittersPType = PCanonicalArray(keyPType)
+          val splittersWasDuplicatedPType = PCanonicalArray(PBooleanRequired)
+
+          val paddedSplittersAddr = cb.newLocal[Long]("stream_dist_splitters_addr", paddedSplittersPType.allocate(region, paddedSplittersSize))
+          cb += paddedSplittersPType.stagedInitialize(paddedSplittersAddr, paddedSplittersSize)
+
+          val splittersWasDuplicatedLength = paddedSplittersSize
+          val splittersWasDuplicatedAddr = cb.newLocal[Long]("stream_dist_dupe_splitters_addr", splittersWasDuplicatedPType.allocate(region, splittersWasDuplicatedLength))
+          cb += splittersWasDuplicatedPType.stagedInitialize(splittersWasDuplicatedAddr, splittersWasDuplicatedLength)
+          val splitters: SIndexableValue = new SIndexablePointerCode(SIndexablePointer(paddedSplittersPType), paddedSplittersAddr).memoize(cb, "stream_distribute_splitters_deduplicated") // last element is duplicated, otherwise this is sorted without duplicates.
+
+          val uniqueSplittersIdx = cb.newLocal[Int]("unique_splitters_idx", 0)
+          val requestedSplittersIdx = cb.newLocal[Int]("stream_distribute_splitters_index")
+          val lastKeySeen = cb.emb.newEmitLocal("stream_distribute_last_seen", keyType, false)
+
+          cb.forLoop(cb.assign(requestedSplittersIdx, 0), requestedSplittersIdx < requestedSplittersVal.loadLength(), cb.assign(requestedSplittersIdx, requestedSplittersIdx + 1), {
+            val currentSplitter = requestedSplittersVal.loadElement(cb, requestedSplittersIdx).memoize(cb, "stream_distribute_current_splitter")
+            cb.ifx(requestedSplittersIdx ceq 0, {
+              paddedSplittersPType.elementType.storeAtAddress(cb, paddedSplittersPType.loadElement(paddedSplittersAddr, paddedSplittersSize, 0), region, currentSplitter.get(cb), false)
+              splittersWasDuplicatedPType.elementType.storeAtAddress(cb, splittersWasDuplicatedPType.loadElement(splittersWasDuplicatedAddr, splittersWasDuplicatedLength, uniqueSplittersIdx), region, new SBooleanCode(false), false)
+              cb.assign(uniqueSplittersIdx, uniqueSplittersIdx + 1)
+            }, {
+              cb.ifx(!equal(cb, lastKeySeen, currentSplitter), {
+                // write to pos in splitters
+                paddedSplittersPType.elementType.storeAtAddress(cb, paddedSplittersPType.loadElement(paddedSplittersAddr, paddedSplittersSize, uniqueSplittersIdx), region, currentSplitter.get(cb), false)
+                splittersWasDuplicatedPType.elementType.storeAtAddress(cb, splittersWasDuplicatedPType.loadElement(splittersWasDuplicatedAddr, splittersWasDuplicatedLength, uniqueSplittersIdx), region, new SBooleanCode(false), false)
+                cb.assign(uniqueSplittersIdx, uniqueSplittersIdx + 1)
+              }, {
+                splittersWasDuplicatedPType.elementType.storeAtAddress(cb, splittersWasDuplicatedPType.loadElement(splittersWasDuplicatedAddr, splittersWasDuplicatedLength, uniqueSplittersIdx - 1), region, new SBooleanCode(true), false)
+              })
+            })
+            cb.assign(lastKeySeen, currentSplitter)
+          })
+
+          val numUniqueSplitters = cb.newLocal[Int]("stream_distribute_num_unique_splitters", uniqueSplittersIdx)
+
+          // Pad out the rest of the splitters array so tree later is balanced.
+          cb.forLoop({}, uniqueSplittersIdx < paddedSplittersSize, cb.assign(uniqueSplittersIdx, uniqueSplittersIdx + 1), {
+            paddedSplittersPType.elementType.storeAtAddress(cb, paddedSplittersPType.loadElement(paddedSplittersAddr, paddedSplittersSize, uniqueSplittersIdx), region, lastKeySeen.get(cb), false)
+          })
+
+
+          val splitterWasDuplicated = new SIndexablePointerCode(SIndexablePointer(splittersWasDuplicatedPType), splittersWasDuplicatedAddr).memoize(cb, "stream_distrib_was_duplicated") // Same length as splitters, but full of booleans of whether it was initially duplicated.
+
+          // Without identity buckets you'd have numUniqueSplitters + 1 buckets, but we have to add an extra for each identity bucket.
+          val numFilesToWrite = cb.newLocal[Int]("stream_dist_num_files_to_write", 1)
+          cb.forLoop(cb.assign(uniqueSplittersIdx, 0), uniqueSplittersIdx < numUniqueSplitters, cb.assign(uniqueSplittersIdx, uniqueSplittersIdx + 1), {
+            cb.assign(numFilesToWrite, numFilesToWrite + 1 + splitterWasDuplicated.loadElement(cb, uniqueSplittersIdx).get(cb).asBoolean.boolCode(cb).toI)
+          })
+
+          val treePType = paddedSplittersPType
+          val treeAddr = cb.newLocal[Long]("stream_dist_tree_addr", treePType.allocate(region, paddedSplittersSize))
+          cb += treePType.stagedInitialize(treeAddr, paddedSplittersSize)
+
+          /*
+          Walk through the array one level of the tree at a time, filling in the tree as you go to get a breadth
+          first traversal of the tree.
+         */
+          val currentHeight = cb.newLocal[Int]("stream_dist_current_height")
+          val treeFillingIndex = cb.newLocal[Int]("stream_dist_tree_filling_idx", 1)
+          cb.forLoop(cb.assign(currentHeight, treeHeight - 1), currentHeight >= 0, cb.assign(currentHeight, currentHeight - 1), {
+            val startingPoint: Value[Int] = cb.newLocal[Int]("stream_dist_starting_point", (const(1) << currentHeight) - 1)
+            val inner = cb.newLocal[Int]("stream_dist_tree_inner")
+            cb.forLoop(cb.assign(inner, 0), inner < (const(1) << (treeHeight - 1 - currentHeight)), cb.assign(inner, inner + 1), {
+              val lookupIndex = cb.newLocal[Int]("temp2", startingPoint + inner * (const(1) << (currentHeight + 1)))
+              val elementLoaded = splitters.loadElement(cb, lookupIndex).get(cb).memoize(cb, "temp")
+              keyPType.storeAtAddress(cb, treePType.loadElement(treeAddr, treeFillingIndex), region,
+                elementLoaded, false)
+              cb.assign(treeFillingIndex, treeFillingIndex + 1)
+            })
+          })
+          // 0th element is garbage, tree elements start at idx 1.
+          val tree: SIndexableValue = new SIndexablePointerCode(SIndexablePointer(treePType), treeAddr).memoize(cb, "stream_dist_tree")
+
+          val outputBuffers = cb.newLocal[Array[OutputBuffer]]("stream_dist_output_buffers", Code.newArray[OutputBuffer](numFilesToWrite))
+          val numElementsPerFile = cb.newLocal[Array[Int]]("stream_dist_elements_per_file", Code.newArray[Int](numFilesToWrite))
+          val fileArrayIdx = cb.newLocal[Int]("stream_dist_file_array_idx")
+
+          def makeFileName(fileIdx: Code[Int]): Code[String] = {
+            pathVal.asString.loadString() concat const("/sorted_part_") concat (fileIdx.toS)
+          }
+
+          cb.forLoop(cb.assign(fileArrayIdx, 0), fileArrayIdx < numFilesToWrite, cb.assign(fileArrayIdx, fileArrayIdx + 1), {
+            val ob = cb.newLocal[OutputBuffer]("stream_dist_file_ob")
+            val fileName = cb.newLocal[String]("file_to_write", makeFileName(fileArrayIdx))
+            cb.assign(ob, spec.buildCodeOutputBuffer(mb.create(fileName)))
+            cb += outputBuffers.update(fileArrayIdx, ob)
+            cb += numElementsPerFile.update(fileArrayIdx, 0)
+          })
+
+          // The element classifying algorithm acts as though there are identity buckets for every splitter. We only use identity buckets for elements that repeat
+          // in splitters list. Since We don't want many empty files, we need to make an array mapping output buckets to files.
+          val fileMappingType = PCanonicalArray(PInt32Required)
+          val fileMappingAddr = cb.newLocal("stream_dist_file_map_addr", fileMappingType.allocate(region, numberOfBuckets))
+          cb += fileMappingType.stagedInitialize(fileMappingAddr, numberOfBuckets)
+
+          val bucketIdx = cb.newLocal[Int]("stream_dist_bucket_idx")
+          val currentFileToMapTo = cb.newLocal[Int]("stream_dist_mapping_cur_storage", 0)
+          def destFileSCode() = new SInt32Code((currentFileToMapTo >= numFilesToWrite).mux(numFilesToWrite - 1, currentFileToMapTo))
+          cb.forLoop(cb.assign(bucketIdx, 0), bucketIdx < numberOfBuckets, cb.assign(bucketIdx, bucketIdx + 2), {
+            fileMappingType.elementType.storeAtAddress(cb, fileMappingType.loadElement(fileMappingAddr, numberOfBuckets, bucketIdx), region, destFileSCode(), false)
+            cb.assign(currentFileToMapTo, currentFileToMapTo + splitterWasDuplicated.loadElement(cb, bucketIdx / 2).get(cb).asBoolean.boolCode(cb).toI)
+            fileMappingType.elementType.storeAtAddress(cb, fileMappingType.loadElement(fileMappingAddr, numberOfBuckets, bucketIdx + 1), region, destFileSCode(), false)
+            cb.assign(currentFileToMapTo, currentFileToMapTo + 1)
+          })
+
+          val fileMapping = new SIndexablePointerCode(SIndexablePointer(fileMappingType), fileMappingAddr).memoize(cb, "stream_dist_file_map")
+
+          val encoder = spec.encodedType.buildEncoder(childStream.st.elementType, cb.emb.ecb)
+          childStream.producer.memoryManagedConsume(region, cb) { cb =>
+            val b = cb.newLocal[Int]("stream_dist_b_i", 1)
+            val current = mb.newEmitField("stream_dist_current", childStream.st.elementEmitType)
+            cb.assign(current, childStream.producer.element)
+
+            val r = cb.newLocal[Int]("stream_dist_r")
+            cb.forLoop(cb.assign(r, 0), r < treeHeight, cb.assign(r, r + 1), {
+              val treeAtB = tree.loadElement(cb, b).memoize(cb, "stream_dist_tree_b")
+              cb.assign(b, const(2) * b + lessThan(cb, treeAtB, current).toI)
+            })
+            cb.assign(b, const(2) * b + 1 - lessThan(cb, current, splitters.loadElement(cb, b - numberOfBuckets / 2).memoize(cb, "stream_dist_splitter_compare")).toI)
+
+            val fileToUse = cb.newLocal[Int]("stream_dist_index_of_file_to_write", fileMapping.loadElement(cb, b - numberOfBuckets).get(cb).asInt.intCode(cb))
+
+            val ob = cb.newLocal[OutputBuffer]("outputBuffer_to_write", outputBuffers(fileToUse))
+
+            cb += ob.writeByte(1.asInstanceOf[Byte])
+            encoder(cb, current.get(cb), ob)
+            cb += numElementsPerFile.update(fileToUse, numElementsPerFile(fileToUse) + 1)
+          }
+
+          cb.forLoop(cb.assign(fileArrayIdx, 0), fileArrayIdx < numFilesToWrite, cb.assign(fileArrayIdx, fileArrayIdx + 1), {
+            val ob = cb.newLocal[OutputBuffer]("stream_dist_output_buffer_to_clean_up", outputBuffers(fileArrayIdx))
+            cb += ob.writeByte(0.asInstanceOf[Byte])
+            cb += ob.invoke[Unit]("close")
+          })
+
+          val intervalType = PCanonicalInterval(keyPType.setRequired(false), true)
+          val returnType = PCanonicalArray(PCanonicalStruct(("interval", intervalType), ("fileName", PCanonicalStringRequired), ("numElements", PInt32Required)), true)
+
+          val (pushElement, finisher) = returnType.constructFromFunctions(cb, region, numFilesToWrite, false)
+
+          val stackStructType = new SStackStruct(ir.typ.asInstanceOf[TArray].elementType.asInstanceOf[TStruct], IndexedSeq(
+            EmitType(intervalType.sType, true),
+            EmitType(SJavaString, true),
+            EmitType(SInt32, true)
+          ))
+
+          // Add first
+          val firstInterval = intervalType.constructFromCodes(cb, region,
+            EmitCode.missing(cb.emb, splitters.st.elementType),
+            EmitCode.fromI(cb.emb)(cb => splitters.loadElement(cb, 0)),
+            EmitCode.present(cb.emb, primitive(false)),
+            EmitCode.present(cb.emb,  new SBooleanCode(!splitterWasDuplicated.loadElement(cb, 0).get(cb).asBoolean.boolCode(cb)))
+          )
+
+          pushElement(cb, IEmitCode.present(cb, new SStackStructCode(stackStructType, IndexedSeq(
+            EmitCode.present(cb.emb, firstInterval),
+            EmitCode.present(cb.emb, SJavaString.construct(makeFileName(0))),
+            EmitCode.present(cb.emb, primitive(numElementsPerFile(0)))
+          ))))
+
+          cb.forLoop({cb.assign(uniqueSplittersIdx, 0); cb.assign(fileArrayIdx, 1) }, uniqueSplittersIdx < numUniqueSplitters, cb.assign(uniqueSplittersIdx, uniqueSplittersIdx + 1), {
+            cb.ifx(uniqueSplittersIdx cne 0, {
+              val intervalFromLastToThis = intervalType.constructFromCodes(cb, region,
+                EmitCode.fromI(cb.emb)(cb => splitters.loadElement(cb, uniqueSplittersIdx - 1)),
+                EmitCode.fromI(cb.emb)(cb => splitters.loadElement(cb, uniqueSplittersIdx)),
+                EmitCode.present(cb.emb, primitive(false)),
+                EmitCode.present(cb.emb, primitive(!splitterWasDuplicated.loadElement(cb, uniqueSplittersIdx).get(cb).asBoolean.boolCode(cb)))
+              )
+
+              pushElement(cb, IEmitCode.present(cb, new SStackStructCode(stackStructType, IndexedSeq(
+                EmitCode.present(cb.emb, intervalFromLastToThis),
+                EmitCode.present(cb.emb, SJavaString.construct(makeFileName(fileArrayIdx))),
+                EmitCode.present(cb.emb, primitive(numElementsPerFile(fileArrayIdx)))
+              ))))
+
+              cb.assign(fileArrayIdx, fileArrayIdx + 1)
+            })
+
+            // Now, maybe have to make an identity bucket.
+            cb.ifx(splitterWasDuplicated.loadElement(cb, uniqueSplittersIdx).get(cb).asBoolean.boolCode(cb), {
+              val identityInterval = intervalType.constructFromCodes(cb, region,
+                EmitCode.fromI(cb.emb)(cb => splitters.loadElement(cb, uniqueSplittersIdx)),
+                EmitCode.fromI(cb.emb)(cb => splitters.loadElement(cb, uniqueSplittersIdx)),
+                EmitCode.present(cb.emb, primitive(true)),
+                EmitCode.present(cb.emb, primitive(true))
+              )
+
+              pushElement(cb, IEmitCode.present(cb, new SStackStructCode(stackStructType, IndexedSeq(
+                EmitCode.present(cb.emb, identityInterval),
+                EmitCode.present(cb.emb, SJavaString.construct(makeFileName(fileArrayIdx))),
+                EmitCode.present(cb.emb, primitive(numElementsPerFile(fileArrayIdx)))
+              ))))
+
+              cb.assign(fileArrayIdx, fileArrayIdx + 1)
+            })
+          })
+
+          // Add last
+          val lastInterval = intervalType.constructFromCodes(cb, region,
+            EmitCode.fromI(cb.emb)(cb => splitters.loadElement(cb, uniqueSplittersIdx - 1)),
+            EmitCode.missing(cb.emb, splitters.st.elementType),
+            EmitCode.present(cb.emb, primitive(false)),
+            EmitCode.present(cb.emb, primitive(false))
+          )
+
+          pushElement(cb, IEmitCode.present(cb, new SStackStructCode(stackStructType, IndexedSeq(
+            EmitCode.present(cb.emb, lastInterval),
+            EmitCode.present(cb.emb, SJavaString.construct(makeFileName(fileArrayIdx))),
+            EmitCode.present(cb.emb,  primitive(numElementsPerFile(fileArrayIdx)))
+          ))))
+
+          finisher(cb)
+        }
+      }
+    }
+  }
+
+}
